Getters instead of magic __get in Hal\Entity
In the `Hal\Collection` class the magic `__get` for getting properties has been deprecated. To get `$collection` the `getCollection` method has to be called instead.
Would it not be better to also use similar solution for the `Hal\Entity` class. So `getEntity` and `getId` methods instead of the magic `__get`?